Lock omniauth-oauth2 to version 1.3.1
omniauth-oauth2 is the oauth library that omniauth plugins use under the
hood. Version 1.4.0 introduced a bug causing most requests to fail,
which can be found here: https://github.com/intridea/omniauth-oauth2/issues/81
It should be fixed in future versions, at which point this change can be
reverted. But for now, this gets registrar working again.

Confirmed to fix the problem (passed testing)