Add minimum version restraints to omniauth and omniauth-google-oauth2
Recently on several projects, when running `bundle update` we noticed
that all oauth* dependencies were getting severely downgraded. The only
thing I can think of is that since we don't have a minimum version lock
on these dependencies, bundler has decided to do whatever it wants with
the versions.

Confirmed that this worked on a couple of our repos.